定时任务表pojo生成添加

+++ novel.manage/src/main/java/com/novel/manage/controller/TestController.java
@@ -1,12 +1,14 @@
 package com.novel.manage.controller;
 
-import com.novel.manage.scheduler.TaskManager;
-import com.novel.manage.scheduler.UtiQuartzConfig;
+ 
 import org.quartz.SchedulerException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+
+import com.novel.common.pojo.UtiQuartzConfig;
+import com.novel.manage.scheduler.TaskManager;
 
 /**
  * Created by Admin on 2017/11/12.
@@ -22,7 +24,7 @@ public class TestController {
     @ResponseBody
     public String testAddTask() throws ClassNotFoundException, NoSuchMethodException, InstantiationException, SchedulerException, IllegalAccessException {
         UtiQuartzConfig utiQuartzConfig = new UtiQuartzConfig();
-        utiQuartzConfig.setId(2);
+        utiQuartzConfig.setId((long) 2);
         utiQuartzConfig.setJobCode("autoTest");
         utiQuartzConfig.setConCurrent("1");
         utiQuartzConfig.setCronExpression("0/5 * * * * ?");
